renames subnav components
